code-review-env/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic>=2.11,<3
openai>=1.30.1,<2
openenv-core
httpx==0.27.0
python-multipart==0.0.9
requests>=2.31,<3
python-dotenv==1.0.1
google-generativeai>=0.8.0

